Upgrade to Ubuntu/trusty64 and emacs-24.5

+++ setup.sh
@@ -20,31 +20,29 @@ VAGRANT_HOME=/home/vagrant
 cd $VAGRANT_HOME
 
 # link to our .emacs file
-sudo -u vagrant ln -s /vagrant/jedi-starter.el $VAGRANT_HOME/.emacs
+sudo -u vagrant mkdir -p $VAGRANT_HOME/.emacs.d
+sudo -u vagrant ln -sf /vagrant/jedi-starter.el $VAGRANT_HOME/.emacs.d/init.el
 
 # get emacs
 
 # build dependencies
 $INSTALL build-essential
-$INSTALL libgtk-3-dev
-$INSTALL libgif-dev libxpm-dev
-$INSTALL texinfo
-$INSTALL libtiff4-dev
-
-# download emacs
-TEMPDIR=install-temp
-mkdir $TEMPDIR
+$INSTALL texinfo libtinfo-dev
 
 # compile
-EMACSBASE=emacs-24.3
+EMACSBASE=emacs-24.5
 EMACSPKG="$EMACSBASE.tar.gz"
-curl -XGET -O "http://ftp.gnu.org/gnu/emacs/$EMACSPKG"
-tar xzvf $EMACSPKG
+if [ ! -f /vagrant/archive/$EMACSPKG ]; then
+    mkdir -p /vagrant/archive
+    curl -XGET -O "http://ftp.gnu.org/gnu/emacs/$EMACSPKG"
+    mv -f $EMACSPKG /vagrant/archive/
+fi
+tar xzvf /vagrant/archive/$EMACSPKG
 cd $EMACSBASE
-./configure
+./configure --without-all --with-zlib
 make
 sudo make install
 
 # and cleanup
 cd $VAGRANT_HOME
-rm -rf $TEMPDIR
+rm -rf $EMACSBASE
